+++ brostar_api/settings.py
@@ -324,7 +324,7 @@ if USE_BRO_PRODUCTION:
     BRONHOUDERSPORTAAL_URL = "https://www.bronhouderportaal-bro.nl"
 else:
     BRO_UITGIFTE_SERVICE_URL = "https://int-publiek.broservices.nl"
-    BRONHOUDERSPORTAAL_URL = "https://int.bronhouderportaal-bro.nl"
+    BRONHOUDERSPORTAAL_URL = "https://demo.bronhouderportaal-bro.nl"
 
 if not DEBUG:
     EMAIL_BACKEND = "django.core.mail.backends.smtp.EmailBackend"